Add snapshot functionality

+++ tests/base_testcase.py
@@ -24,7 +24,7 @@ class TestCaseBase:
         self._next_p2p_port = None
         self._next_rest_port = None
 
-    def initialise(self, wd_parent_path: str = None, host: str = '127.0.0.1',
+    def initialise(self, wd_parent_path: str = None, snapshot_path: str = None, host: str = '127.0.0.1',
                    next_p2p_port: int = 4000, next_rest_port: int = 5000) -> None:
         # determine the working directory for testing
         if wd_parent_path:
@@ -37,8 +37,12 @@ class TestCaseBase:
         if os.path.exists(self.wd_path):
             raise Exception(f"path to working directory for testing '{self.wd_path}' already exists!")
 
-        # create working directory
-        os.makedirs(self.wd_path, exist_ok=True)
+        # do we have a snapshot path? if so, copy all the contents
+        if snapshot_path:
+            shutil.copytree(snapshot_path, self.wd_path)
+        else:
+            # create an enpty working directory
+            os.makedirs(self.wd_path, exist_ok=True)
 
         self.host = host
         self.nodes = {}
@@ -47,14 +51,17 @@ class TestCaseBase:
         self._next_p2p_port = next_p2p_port
         self._next_rest_port = next_rest_port
 
-    def cleanup(self) -> None:
+    def cleanup(self, snapshot_path: str = None) -> None:
         for name in self.nodes:
             logger.info(f"stopping node '{name}'")
             node = self.nodes[name]
             node.shutdown()
 
-        # delete working directory
-        shutil.rmtree(self.wd_path)
+        # retain the working directory as snapshot?
+        if snapshot_path:
+            os.rename(self.wd_path, snapshot_path)
+        else:
+            shutil.rmtree(self.wd_path)
 
     def generate_p2p_address(self) -> (str, int):
         with self._mutex:
@@ -128,6 +135,37 @@ class TestCaseBase:
 
             else:
                 keystore = Keystore.create(storage_path, name, f"no-email-provided", 'password')
+
+            # create node and startup services
+            node = Node(keystore, storage_path)
+            node.startup(p2p_address, enable_dor=use_dor, enable_rti=use_rti,
+                         rest_address=rest_address if enable_rest else None,
+                         retain_job_history=retain_job_history)
+
+            self.nodes[name] = node
+            return node
+
+    def resume_node(self, name: str, enable_rest: bool = False, use_dor: bool = True, use_rti: bool = True,
+                    retain_job_history: bool = True) -> Node:
+        if name in self.nodes:
+            return self.nodes[name]
+
+        else:
+            p2p_address = self.generate_p2p_address()
+            rest_address = self.generate_rest_address()
+
+            storage_path = os.path.join(self.wd_path, name)
+            if not os.path.isdir(storage_path):
+                raise RuntimeError(f"no storage path found to resume node at {storage_path}")
+
+            # infer the keystore id
+            keystore_id = None
+            for filename in os.listdir(storage_path):
+                if filename.endswith('.json') and len(filename) == 69:
+                    keystore_id = filename.split('.')[0]
+                    break
+
+            keystore = Keystore.load(storage_path, keystore_id, 'password')
 
             # create node and startup services
             node = Node(keystore, storage_path)
